+++ src/FESpaces/FESpaceFactories.jl
@@ -1,7 +1,8 @@
 
 function FESpace(
   model::DiscreteModel,
-  cell_fe::CellFE;
+  cell_fe::CellFE,
+  cell_conformity::CellConformity=CellConformity(cell_fe);
   labels = get_face_labeling(model),
   dirichlet_tags=Int[],
   dirichlet_masks=nothing,
@@ -12,9 +13,21 @@ function FESpace(
   The number of cells provided in the `cell_fe` argument ($(cell_fe.num_cells) cells)
   does not match the number of cells ($(num_cells(model)) cells) in the provided DiscreteModel.
   """
-
   trian = Triangulation(model)
+  _vector_type = _get_vector_type(vector_type,cell_fe,trian)
+  F = _ConformingFESpace(
+      _vector_type,
+      model,
+      labels,
+      cell_fe,
+      cell_conformity,
+      dirichlet_tags,
+      dirichlet_masks)
+  V = _add_constraint(F,cell_fe.max_order,constraint)
+  V
+end
 
+function _get_vector_type(vector_type,cell_fe,trian)
   if vector_type == nothing
     cell_shapefuns, cell_dof_basis = compute_cell_space(cell_fe,trian)
     T = get_dof_value_type(cell_shapefuns,cell_dof_basis)
@@ -25,22 +38,7 @@ function FESpace(
     """
     _vector_type = vector_type
   end
-
-  if CellConformity(cell_fe).conformity == L2Conformity() && dirichlet_tags == Int[]
-    F = _DiscontinuousFESpace(_vector_type,trian,cell_fe)
-  else
-    F = _ConformingFESpace(
-      _vector_type,
-      model,
-      labels,
-      cell_fe,
-      dirichlet_tags,
-      dirichlet_masks)
-  end
-
-  V = _add_constraint(F,cell_fe.max_order,constraint)
-
-  V
+  _vector_type
 end
 
 function _add_constraint(F,order,constraint)
@@ -61,10 +59,24 @@ end
 
 function FESpace(
   model::RestrictedDiscreteModel,
-  cell_fe::CellFE;
+  cell_reffe::AbstractArray{<:ReferenceFE};
+  conformity=nothing,
   constraint=nothing,kwargs...)
   model_portion = model.model
+  cell_fe = CellFE(model,cell_reffe,conformity)
   V_portion = FESpace(model_portion,cell_fe;constraint=nothing,kwargs...)
+  F = ExtendedFESpace(V_portion,model)
+  V = _add_constraint(F,cell_fe.max_order,constraint)
+  V
+end
+
+function FESpace(
+  model::RestrictedDiscreteModel,
+  cell_fe::CellFE,
+  cell_conformity::CellConformity=CellConformity(cell_fe);
+  constraint=nothing,kwargs...)
+  model_portion = model.model
+  V_portion = FESpace(model_portion,cell_fe,cell_conformity;constraint=nothing,kwargs...)
   F = ExtendedFESpace(V_portion,model)
   V = _add_constraint(F,cell_fe.max_order,constraint)
   V
@@ -74,13 +86,27 @@ function FESpace(
   model::DiscreteModel,
   cell_reffe::AbstractArray{<:ReferenceFE};
   conformity=nothing,
-  kwargs...)
+  labels = get_face_labeling(model),
+  dirichlet_tags=Int[],
+  dirichlet_masks=nothing,
+  constraint=nothing,
+  vector_type::Union{Nothing,Type}=nothing)
+
   trian = Triangulation(model)
-  reffe = testitem(cell_reffe)
-  conf = Conformity(reffe,conformity)
-  cell_conformity = CellConformity(cell_reffe,conf)
-  cell_fe = CellFE(model,cell_reffe,cell_conformity)
-  FESpace(model,cell_fe;kwargs...)
+  cell_fe = CellFE(model,cell_reffe,conformity)
+  _vector_type = _get_vector_type(vector_type,cell_fe,trian)
+  if conformity in (L2Conformity(),:L2) && dirichlet_tags == Int[]
+    F = _DiscontinuousFESpace(_vector_type,trian,cell_fe)
+    V = _add_constraint(F,cell_fe.max_order,constraint)
+  else
+    V = FESpace(model,cell_fe;
+                labels=labels,
+                dirichlet_tags=dirichlet_tags,
+                dirichlet_masks=dirichlet_masks,
+                constraint=constraint,
+                vector_type=_vector_type)
+  end
+  return V
 end